requirements.txt
bitly-api-py3
colorama